separate imageUpload into controller file
